Renamed method from Renderer::renderPlayer to Renderer::renderEntity, changed hardcoding to softcoding for some values, like window size, map elements size and entity sizes (now size is relative to map elements's size : 0.5 is 50% of the size, 1 is 100%)

+++ include/entity/npc/Npc.h
@@ -6,7 +6,7 @@
 class Npc : public Renderable
 {
 public:
-    Npc(char* name, int x, int y, int width, int height, char* sprite);
+    Npc(char* name, int x, int y, float width, float height, char* sprite);
     virtual void update() = 0;
 
 };
